Explicit “renderToElement” API
The previous Node rendering tests were fairly brittle due to them faking
out a Morph and overriding Ember.View’s `_morph` property. Indeed, this
was just a short-term hack and the actual FastBoot API departed from it.

Because of that, the recent outlet refactor broke the FastBoot
implementation because it changed the application’s root view to be
tag-less, breaking the assumption that the root view would always have
an element.

This commit makes explicit the API for rendering an an application’s
view hierarchy into an element detached from the document’s body.

In particular, the implementation of `appendTo` relied on jQuery to
normalize selectors into elements. In FastBoot, jQuery is not available.
Instead, we introduce a new `renderToElement` method that creates a new,
detached `body` element and instructs Ember to render the view inside
that element. Once rendering is complete, the `body` element is
returned.

Hopefully by making these APIs more explicit, we can ensure future
commits do not break FastBoot rendering.